# speed up registry Each method