@mihasya 2018-09-11
Am traveling, unlikely to look soon. Always appreciate a patch with
benchmarks included, making calendar reminder to look when back. Thank you!

Alternatively, @wadey @rcrowley you around? :D

On Tue, Sep 11, 2018 at 09:24 Ingo Oeser <notifications@github.com> wrote:

> Creating a map entry with a string type as key leads to a copy of that
> string in Go. Iterating via Each creates a lot of these entries,
> especially while reporting lots of values to a remote metrics system.
>
> Our use case is modelled with HugeRegistry and the below benchmarks show
> improvements there to to the removed memory allocation.
>
> name                  old time/op    new time/op    delta
> Counter-4               9.83ns ± 0%   10.03ns ± 1%   +2.03%          (p=0.016 n=4+5)
> GuageFloat64-4          62.7ns ± 4%    60.6ns ± 1%   -3.38%          (p=0.008 n=5+5)
> Histogram-4              117ns ± 3%     129ns ± 4%  +10.26%          (p=0.008 n=5+5)
> Registry-4               760ns ± 2%     232ns ± 2%  -69.52%          (p=0.008 n=5+5)
> HugeRegistry-4          2.58ms ± 3%    0.82ms ± 1%  -68.10%          (p=0.008 n=5+5)
> UniformSample257-4       108ns ± 2%     123ns ± 5%  +13.10%          (p=0.008 n=5+5)
> UniformSample514-4       106ns ± 2%     118ns ± 5%  +10.90%          (p=0.008 n=5+5)
> UniformSample1028-4      106ns ± 1%     119ns ± 2%  +12.43%          (p=0.008 n=5+5)
>
> name                  old alloc/op   new alloc/op   delta
> Registry-4                336B ± 0%       32B ± 0%  -90.48%          (p=0.008 n=5+5)
> HugeRegistry-4           598kB ± 0%     328kB ± 0%     ~             (p=0.079 n=4+5)
>
> name                  old allocs/op  new allocs/op  delta
> Registry-4                2.00 ± 0%      1.00 ± 0%  -50.00%          (p=0.008 n=5+5)
> HugeRegistry-4            5.00 ± 0%      3.00 ± 0%  -40.00%          (p=0.008 n=5+5)
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/rcrowley/go-metrics/pull/242
> Commit Summary
>
>    - speed up registry Each method
>
> File Changes
>
>    - *M* registry.go
>    <https://github.com/rcrowley/go-metrics/pull/242/files#diff-0> (225)
>    - *M* registry_test.go
>    <https://github.com/rcrowley/go-metrics/pull/242/files#diff-1> (251)
>
> Patch Links:
>
>    - https://github.com/rcrowley/go-metrics/pull/242.patch
>    - https://github.com/rcrowley/go-metrics/pull/242.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rcrowley/go-metrics/pull/242>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AABYRFtRoxzCY9_yWvdgPflm5mgJgCRbks5uZ2VAgaJpZM4Wivhf>
> .
>
-- 
Sent from Russia, with love. But also from my phone.