@mihasya 2018-10-15
@nightlyone hey there! Did you intend this PR as a starting point? I just took a look, and it appears that the PR deletes a whole bunch of interfaces and implementations, including `PrefixedRegistry`, without replacing them. Performance optimizations are always welcome, but I can't merge something that removes functionality.