@nightlyone 2018-10-15
@mihasya Sorry, that looked like some merge-foo. 

So I redid the changes. Please take a look again.