@nightlyone 2018-11-16
@mihasya wanna take a look again?

Please note that this change now conflicts with pull-request #247 
I can rebase my change again on short notice and redo the benchmarks, if you want to merge #247 first. 

It would really be great to get those changes merged. We are still using a fork including this change at the moment in production and I don't want to keep doing that forever :-)