@mihasya 2019-01-18
Sorry about the silence - we've been slowly discussing moving this repo onto an actual GH org, which would allow us to actually manage it sanely. Because it's under Richard's personal org, only he is able to modify integrations. We want to make sure we don't do the move rashly and break everybody. The expectation is that a moved repo will redirect automatically, but I haven't had the time to research/verify that the git tooling will work transparently. 

Basically, we are the Ents in Lord of the Rings, and we've only just said good morning 🤦‍♂️