@nightlyone 2019-04-15
And another month passed. @mihasya / @wadey / @rcrowley are you still maintaining this library or should there be a fork started which collects all the fixes/improvements so people can move on? More than half a year to merge a PR makes it a bit difficult to usefully contribute for me.