@mihasya 2019-04-15
We deserve all the ire, I'm really sorry this is dragging on so much. The problem is that it's clear the next logical thing is to move this repo to an org so that we can more easily manage access etc. I was planning to perform the move, then fix all the CI/CD hookups, then go through and merge everything. Unfortunately the move is a sufficiently large undertaking that I haven't been able to find the time to even plan it. I will get with @rcrowley in the near future to just fix the repo in place again, work through this logjam of pull requests, then figure out the move story.