@temoto 2019-05-04
@mihasya is it possible to subscribe to news about repo move event?